+++ dpcomp/include/plier/dialect/plier/PlierOps.td
@@ -202,55 +202,6 @@ def ExhaustIterOp : Plier_Op<"exhaust_iter", [NoSideEffect]> {
                              : $count)>];
 }
 
-def EnforceShapeOp : Plier_Op<"enforce_shape"> {
-  let arguments = (ins AnyRankedTensor : $value, Variadic<Index> : $sizes);
-
-  let results = (outs AnyRankedTensor : $result);
-
-  let builders = [OpBuilder<(ins "::mlir::Value"
-                             : $value, "::mlir::ValueRange"
-                             : $shape)>];
-
-  let hasFolder = 1;
-  let hasCanonicalizer = 1;
-}
-
-def ParallelOp : Plier_Op<"parallel", [
-  AttrSizedOperandSegments, DeclareOpInterfaceMethods<LoopLikeOpInterface>,
-  SingleBlockImplicitTerminator<"plier::YieldOp">, RecursiveSideEffects
-]> {
-
-  let arguments = (ins Variadic<Index>:$lowerBounds,
-                         Variadic<Index>:$upperBounds,
-                         Variadic<Index>:$steps);
-  let regions = (region SizedRegion<1> : $region);
-
-  let skipDefaultBuilders = 1;
-  let builders = [OpBuilder<(
-      ins "::mlir::ValueRange"
-      : $lowerBounds, "::mlir::ValueRange"
-      : $upperBounds, "::mlir::ValueRange"
-      : $steps,
-        CArg<"::mlir::function_ref<void(::mlir::OpBuilder &, ::mlir::Location, "
-             "::mlir::ValueRange, ::mlir::ValueRange, ::mlir::Value)>",
-             "nullptr">)>];
-
-    let extraClassDeclaration = [{
-        unsigned getNumLoops() { return steps().size();
-}
-}];
-}
-
-def YieldOp : Plier_Op<"yield", [
-  NoSideEffect, ReturnLike, Terminator, ParentOneOf<["ParallelOp"]>
-]> {
-  let arguments = (ins Variadic<AnyType> : $results);
-  let builders = [OpBuilder<(ins), [{/* nothing to do */}]>];
-  // Override default verifier (defined in SCF_Op), no custom verification
-  // needed.
-  let verifier = ? ;
-}
-
 def UndefOp : Plier_Op<"undef", [NoSideEffect]> {
   let results = (outs AnyType);
 }
@@ -274,183 +225,5 @@ def SliceGetItemOp : Plier_Op<"slice_getitem", [NoSideEffect]> {
 
   let hasCanonicalizer = 1;
 }
-
-def RetainOp : Plier_Op<"retain", [ViewLikeOpInterface]> {
-  let arguments = (ins AnyMemRef : $source);
-
-  let results = (outs AnyMemRef);
-
-  let builders = [OpBuilder<(ins "::mlir::Value" : $value)>];
-
-  let extraClassDeclaration = [{
-      ::mlir::Value getViewSource() { return source(); }
-  }];
-}
-
-def ChangeLayoutOp : Plier_Op<"change_layout", [ViewLikeOpInterface, NoSideEffect]> {
-  let arguments = (ins AnyMemRef : $source);
-
-  let results = (outs AnyMemRef);
-  let hasFolder = 1;
-  let hasCanonicalizer = 1;
-
-  let extraClassDeclaration = [{
-      ::mlir::Value getViewSource() { return source(); }
-  }];
-}
-
-def SignCastOp : Plier_Op<"sign_cast", [NoSideEffect]> {
-  let arguments = (ins AnyType : $value);
-
-  let results = (outs AnyType);
-  let hasFolder = 1;
-  let hasCanonicalizer = 1;
-}
-
-def ReduceRankOp
-    : Plier_Op<"reduce_rank", [ViewLikeOpInterface, NoSideEffect]> {
-  let arguments = (ins
-                   AnyTypeOf<[AnyMemRef, AnyRankedTensor]>:$source,
-                   I32ArrayAttr:$mapping);
-
-  let results = (outs AnyTypeOf<[AnyMemRef, AnyRankedTensor]>);
-  let hasFolder = 1;
-  let hasCanonicalizer = 1;
-
-  let builders = [
-    OpBuilder<(ins "::mlir::Value" : $src,
-                   "::mlir::ArrayRef<int32_t>" : $mapping)>
-  ];
-
-  let extraClassDeclaration = [{
-      ::mlir::Value getViewSource() { return source(); }
-      ::llvm::SmallVector<int32_t> getMapping();
-  }];
-}
-
-def ExtractMemrefMetadataOp
-    : Plier_Op<"extract_memref_metadata", [NoSideEffect]> {
-  let arguments = (ins AnyMemRef : $source, IndexAttr : $dimIndex);
-
-  let results = (outs Index : $result);
-  let hasFolder = 1;
-
-  let builders = [
-    OpBuilder<(ins "::mlir::Value" : $src,
-                 "int64_t" : $dim)>,
-    OpBuilder<(ins "::mlir::Value" : $src)>
-  ];
-}
-
-def CreateGpuStreamOp : Plier_Op<"create_gpu_stream", [NoSideEffect]> {
-  let results = (outs Plier_OpaqueType : $result);
-
-  let hasCanonicalizer = 1;
-
-  let builders = [OpBuilder<(ins)>];
-}
-
-def DestroyGpuStreamOp : Plier_Op<"destroy_gpu_stream"> {
-  let arguments = (ins Plier_OpaqueType : $source);
-}
-
-def LoadGpuModuleOp : Plier_Op<"load_gpu_module", [NoSideEffect]> {
-  let arguments = (ins Plier_OpaqueType : $stream, SymbolRefAttr : $module);
-  let results = (outs Plier_OpaqueType : $result);
-
-  let hasCanonicalizer = 1;
-
-  let builders = [OpBuilder<(ins "::mlir::Value"
-                             : $stream, "::mlir::gpu::GPUModuleOp"
-                             : $module)>];
-}
-
-def DestroyGpuModuleOp : Plier_Op<"destroy_gpu_module"> {
-  let arguments = (ins Plier_OpaqueType : $source);
-}
-
-def GetGpuKernelOp : Plier_Op<"get_gpu_kernel", [NoSideEffect]> {
-  let arguments = (ins Plier_OpaqueType : $module, SymbolRefAttr : $kernel);
-  let results = (outs Plier_OpaqueType : $result);
-
-  let hasCanonicalizer = 1;
-
-  let builders = [OpBuilder<(ins "::mlir::Value"
-                             : $module, "::mlir::gpu::GPUFuncOp"
-                             : $kernel)>];
-}
-
-def LaunchGpuKernelOp
-    : Plier_Op<"launch_gpu_kernel",
-               [GPU_AsyncOpInterface, AttrSizedOperandSegments]> {
-  let arguments = (ins Variadic<GPU_AsyncToken>:$asyncDependencies,
-                     Plier_OpaqueType:$stream,
-                     Plier_OpaqueType:$kernel,
-                     Index:$gridSizeX, Index:$gridSizeY, Index:$gridSizeZ,
-                     Index:$blockSizeX, Index:$blockSizeY, Index:$blockSizeZ,
-                     Variadic<AnyType>:$operands);
-  let results = (outs Optional<GPU_AsyncToken> : $asyncToken);
-
-  let skipDefaultBuilders = 1;
-  let builders = [OpBuilder<(ins "::mlir::Value" : $stream,
-                                 "::mlir::Value" : $kernel,
-                                 "::mlir::gpu::KernelDim3" : $gridSize,
-                                 "::mlir::gpu::KernelDim3" : $blockSize,
-                                 "::mlir::ValueRange" : $kernelOperands)>];
-}
-
-def DestroyGpuKernelOp : Plier_Op<"destroy_gpu_kernel"> {
-  let arguments = (ins Plier_OpaqueType : $source);
-}
-
-def GPUAllocOp
-    : Plier_Op<"gpu_alloc", [GPU_AsyncOpInterface, AttrSizedOperandSegments]> {
-
-  let arguments = (ins Variadic<GPU_AsyncToken>:$asyncDependencies,
-                       Plier_OpaqueType:$stream,
-                       Variadic<Index>:$dynamicSizes,
-                       Variadic<Index>:$symbolOperands);
-  let results = (outs Res<AnyMemRef, "", [MemAlloc]>:$memref,
-                      Optional<GPU_AsyncToken>:$asyncToken);
-
-  let extraClassDeclaration = [{
-    ::mlir::MemRefType getType() { return memref().getType().cast<::mlir::MemRefType>(); }
-  }];
-}
-
-def GPUSuggestBlockSizeOp : Plier_Op<"suggest_block_size",
-                                     [AttrSizedOperandSegments, NoSideEffect]> {
-  let arguments = (ins Optional<Plier_OpaqueType>:$stream,
-                       Optional<Plier_OpaqueType>:$kernel,
-                       OptionalAttr<SymbolRefAttr>:$kernelRef,
-                       Variadic<Index>:$gridSize);
-
-  let results = (outs Variadic<Index>);
-
-  let builders = [OpBuilder<(ins "::llvm::Optional<::mlir::Value>" : $stream,
-                                 "::mlir::OpFoldResult" : $kernel,
-                                 "::mlir::ValueRange" : $gridSize)>];
-
-  let extraClassDeclaration = [{
-      /// The name of the kernel's containing module.
-      ::mlir::StringAttr getKernelModuleName();
-
-      /// The name of the kernel.
-      ::mlir::StringAttr getKernelName();
-  }];
-}
-
-def ForceCopyOp : Plier_Op<"force_copy", [ViewLikeOpInterface]> {
-  let arguments = (ins AnyRankedTensor : $source);
-
-  let results = (outs AnyRankedTensor);
-
-  let builders = [OpBuilder<(ins "::mlir::Value" : $value)>];
-
-  let extraClassDeclaration = [{
-      ::mlir::Value getViewSource() { return source(); }
-  }];
-}
-
 
 #endif // PLIER_OPS
