+++ dpcomp/include/plier/dialect/plier_util/CMakeLists.txt
@@ -1,0 +1,12 @@
+include_directories(${MLIR_INCLUDE_DIRS})
+set(dialect PlierUtilOps)
+set(dialect_namespace plier_util)
+set(LLVM_TARGET_DEFINITIONS ${dialect}.td)
+mlir_tablegen(${dialect}Enums.h.inc -gen-enum-decls)
+mlir_tablegen(${dialect}Enums.cpp.inc -gen-enum-defs)
+mlir_tablegen(${dialect}.h.inc -gen-op-decls)
+mlir_tablegen(${dialect}.cpp.inc -gen-op-defs)
+mlir_tablegen(${dialect}Dialect.h.inc -gen-dialect-decls -dialect=${dialect_namespace})
+mlir_tablegen(${dialect}Dialect.cpp.inc -gen-dialect-defs -dialect=${dialect_namespace})
+add_public_tablegen_target(MLIR${dialect}IncGen)
+add_dependencies(mlir-headers MLIR${dialect}IncGen)
