+++ numba_dpcomp/mlir_compiler/lib/pipelines/lower_to_llvm.cpp
@@ -53,6 +53,7 @@
 
 #include "plier/compiler/pipeline_registry.hpp"
 #include "plier/dialect/plier/dialect.hpp"
+#include "plier/dialect/plier_util/dialect.hpp"
 #include "plier/transforms/func_utils.hpp"
 #include "plier/transforms/type_conversion.hpp"
 #include "plier/utils.hpp"
