+++ numba_dpcomp/mlir_compiler/lib/pipelines/pre_low_simplifications.cpp
@@ -19,6 +19,7 @@
 
 #include "plier/compiler/pipeline_registry.hpp"
 #include "plier/dialect/plier/dialect.hpp"
+#include "plier/dialect/plier_util/dialect.hpp"
 #include "plier/transforms/type_conversion.hpp"
 
 #include <mlir/Dialect/Arithmetic/IR/Arithmetic.h>
