+++ numba_dpcomp/mlir_compiler/lib/py_linalg_resolver.cpp
@@ -29,6 +29,7 @@
 #include <mlir/Parser.h>
 
 #include "plier/dialect/plier/dialect.hpp"
+#include "plier/dialect/plier_util/dialect.hpp"
 #include "plier/transforms/const_utils.hpp"
 #include "plier/transforms/func_utils.hpp"
 #include "plier/utils.hpp"
